+++ contracts/dao-light/src/state.rs
@@ -1,9 +1,10 @@
-use crate::{Member, Proposal};
-use codec::{Decode, Encode};
 use gstd::{prelude::*, ActorId};
-use scale_info::TypeInfo;
+
+use crate::{Member, Proposal};
 
 #[derive(Debug, Decode, Encode, TypeInfo)]
+#[codec(crate = gstd::codec)]
+#[scale_info(crate = gstd::scale_info)]
 pub enum Role {
     Admin,
     Member,
@@ -11,6 +12,8 @@ pub enum Role {
 }
 
 #[derive(Debug, Decode, Encode, TypeInfo)]
+#[codec(crate = gstd::codec)]
+#[scale_info(crate = gstd::scale_info)]
 pub enum State {
     UserStatus(ActorId),
     AllProposals,
@@ -22,6 +25,8 @@ pub enum State {
 }
 
 #[derive(Debug, Encode, Decode, TypeInfo)]
+#[codec(crate = gstd::codec)]
+#[scale_info(crate = gstd::scale_info)]
 pub enum StateReply {
     UserStatus(Role),
     AllProposals(BTreeMap<u128, Proposal>),