+++ contracts/dao-light/tests/dao_simple_test.rs
@@ -1,5 +1,5 @@
-use codec::Encode;
-use dao_light_io::*;
+use dao_light::io::*;
+use gstd::Encode;
 use gtest::System;
 mod utils;
 use utils::*;