+++ contracts/dao-light/tests/utils.rs
@@ -1,4 +1,4 @@
-use dao_light_io::*;
+use dao_light::io::*;
 use ft_io::*;
 use gtest::{Program, RunResult, System};
 pub const MEMBERS: &[u64] = &[3, 4, 5, 6];
@@ -6,7 +6,7 @@ pub const ZERO_ID: u64 = 0;
 
 pub fn init_fungible_token(sys: &System) {
     sys.init_logger();
-    let ft = Program::from_file(sys, "./target/fungible_token-0.1.0.wasm");
+    let ft = Program::from_file(sys, "./target/fungible_token-0.1.3.wasm");
 
     let res = ft.send(
         MEMBERS[0],