+++ contracts/dao-light/CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2022-11-30
+### Changed
+- Updated `gstd`, `gtest`, `gear-wasm-builder` to the `stable` branch.
+### Removed
+- `dao-light-io` crate.
+
 ## [0.1.0] - 2022-07-27
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/dao-light/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/gear-dapps/dao-light/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/gear-dapps/dao-light/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/gear-dapps/dao-light/compare/57153c6...0.1.0